Bing News Summarizer/requirements.txt
requests
os
openai
python-dotenv
flask
flask_cors
streamlit